Neousys: reorganise stuff from init to own file and rename NeousysWdtBus to NeousysBus
